// sdk/core/learning/LearningState.ts
export enum LearningState {
    CREATED = 'CREATED',
    COLLECTING_RESULTS = 'COLLECTING_RESULTS',
    PATTERN_MINED = 'PATTERN_MINED',
    ANALYZING = 'ANALYZING',
    LEARNING = 'LEARNING',
    GENERATING_IMPROVEMENTS = 'GENERATING_IMPROVEMENTS',
    VALIDATING = 'VALIDATING',
    PROMOTED = 'PROMOTED',
    ARCHIVED = 'ARCHIVED'
}